BackEnd/server/requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
protobuf==3.17.3
Keras==2.4.3
tensorflow==2.5.0
scikit-learn==0.24.2
google-cloud
google-cloud-vision
google-cloud-storage
Pillow
numpy
expertai-nlapi
h5py
gcsfs